feat\!: migrate from setup.py to pure pyproject.toml build system
BREAKING CHANGE: Remove setup.py in favor of modern pyproject.toml configuration

- Add custom build backend (_build_backend.py) for dynamic metadata injection
- Extract build commands to _build_commands.py (PrepareStatic, BuildPy, SdistCommand)
- Create _dynamic_dependencies.py for centralized dependency resolution
- Maintain all existing functionality:
  - LOCK_REQUIREMENTS environment variable support
  - DEV_SOURCE_PATH support for unpinning dependencies
  - DJANGO_ANSIBLE_BASE_BRANCH support for DAB version selection
  - Custom prepare_static command for UI asset downloading
- Configure bumpversion to update version in _dynamic_dependencies.py
- Ensure backwards compatibility with existing development workflows
