+++ pyproject.toml
@@ -1,3 +1,51 @@
+[build-system]
+requires = ["setuptools>=61.0", "wheel"]
+build-backend = "_build_backend"
+backend-path = ["."]
+
+[project]
+name = "galaxy-ng"
+dynamic = ["version", "dependencies"]
+description = "galaxy-ng plugin for the Pulp Project"
+readme = "README.md"
+license = {text = "GPLv2+"}
+authors = [
+    {name = "Red Hat, Inc.", email = "info@ansible.com"}
+]
+requires-python = ">=3.9"
+classifiers = [
+    "License :: OSI Approved :: GNU General Public License v2 or later (GPLv2+)",
+    "Operating System :: POSIX :: Linux", 
+    "Framework :: Django",
+    "Programming Language :: Python",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.8",
+    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+]
+
+[project.urls]
+Homepage = "https://github.com/ansible/galaxy_ng/"
+
+[project.entry-points."pulpcore.plugin"]
+galaxy_ng = "galaxy_ng:default_app_config"
+
+[tool.setuptools]
+py-modules = ["_build_commands", "_dynamic_dependencies"]
+include-package-data = true
+
+[tool.setuptools.packages.find]
+exclude = ["tests", "tests.*"]
+
+[tool.setuptools.cmdclass]
+prepare_static = "_build_commands.PrepareStaticCommand"
+build_py = "_build_commands.BuildPyCommand"
+sdist = "_build_commands.SDistCommand"
+
+# Dynamic version and dependencies metadata is handled by the custom build backend
+# See _build_backend.py, _build_commands.py, and _dynamic_dependencies.py
+
 [tool.black]
 line-length = 100
 target-version = ["py36", "py37"]
@@ -12,7 +60,6 @@ exclude = '''
   | migrations
 )/
 '''
-
 
 [tool.darker]
 src = [
@@ -65,6 +112,9 @@ ignore = [
     "pr_check.sh",
     "lint_requirements.txt",
     "profiles/**",
+    "_build_commands.py",
+    "_dynamic_dependencies.py",
+    "_build_backend.py",
 ]
 
 [tool.coverage.run]
@@ -210,6 +260,6 @@ replace = "__version__ = \"{new_version}\""
 
 [[tool.bumpversion.files]]
 
-filename = "./setup.py"
+filename = "./_dynamic_dependencies.py"
 search = "version = \"{current_version}\""
 replace = "version = \"{new_version}\""
